more hosts and better suggestions
